requirements.txt
torch>=1.13.0
torchvision>=0.14.0
Pillow>=9.0.0
numpy>=1.20.0
gradio>=3.0.0 